removed temporary component PointsListPag
