+++ src/components/adminPanel/ManagePoints/ManagePoints.tsx
@@ -9,8 +9,7 @@ import Typography from '@mui/material/Typography';
 
 import classes from './ManagePoints.module.scss';
 import AdminHeader from '../AdminHeader';
-// import PointsList from '../PointsList';
-import PointsListPag from '../PointsListPag';
+import PointsList from '../PointsList';
 import PointsMap from '../PointsMap';
 
 function a11yProps(index: number) {
@@ -70,8 +69,7 @@ const ManagePoints = () => {
                         </Button>
                     </div>
                     <CustomTabPanel value={value} index={0}>
-                        {/* <PointsList /> */}
-                        <PointsListPag />
+                        <PointsList />
                     </CustomTabPanel>
                     <CustomTabPanel value={value} index={1}>
                         <PointsMap />
