+++ src/components/adminPanel/PointsList/PointsList.module.scss
@@ -15,3 +15,21 @@
 .noPhone {
   color: rgba($color-black, 0.38);
 }
+
+.pagination {
+  // width: 100%;
+  margin-top: 10px;
+  display: flex;
+  justify-content: flex-end;
+
+  & div{
+    display: flex;
+    align-items: flex-start;
+  }
+  
+  p {
+    white-space: nowrap;
+    width: auto;
+    margin: 0 10px;
+  }
+}
