--- src/components/adminPanel/PointsListPag/index.ts
@@ -1,2 +1,0 @@
-import PointsList from './PointsList';
-export default PointsList;
